% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudsolve.R
\name{solveGameQuick}
\alias{solveGameQuick}
\title{Sudoku solver}
\usage{
solveGameQuick(sudoku)
}
\arguments{
\item{sudoku}{the puzzle to be solved}
}
\description{
Solves a given sudoku, first trying human methods (currently just lone and
hidden singles) and then useing a brute force backtracking approach.
}
