% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudsolve.R
\name{solveHuman}
\alias{solveHuman}
\title{Sudoku solver (human methods only) - NOT EXPORTED}
\usage{
solveHuman(sudoku)
}
\arguments{
\item{sudoku}{the puzzle to be solved}
}
\description{
Attempts to solve a sudoku using human methods (currently just lone and
hidden singles)
}
