% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudsolve.R
\name{sudsolve_brute}
\alias{sudsolve_brute}
\title{Sudoku solver (brute force only) - NOT EXPORTED}
\usage{
sudsolve_brute(sudoku)
}
\arguments{
\item{sudoku}{the puzzle to be solved}
}
\description{
Solves a given sudoku using the brute-force backtracing method
}
